requirements.txt
accelerate==0.29.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.23.1
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
datasets==2.18.0
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
filelock==3.13.3
FlagEmbedding==1.2.8
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
llama-index==0.10.15
llama-index-agent-openai==0.1.7
llama-index-cli==0.1.11
llama-index-core==0.10.27
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-anthropic==0.1.9
llama-index-llms-openai==0.1.14
llama-index-multi-modal-llms-openai==0.1.4
llama-index-postprocessor-flag-embedding-reranker==0.1.2
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.13
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.0
llamaindex-py-client==0.1.16
MarkupSafe==2.1.5
marshmallow==3.21.1
minijinja==1.0.16
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.16.2
packaging==24.0
pandas==2.2.1
pendulum==3.0.0
pillow==10.3.0
psutil==5.9.8
pyarrow==15.0.2
pyarrow-hotfix==0.6
pydantic==2.6.4
pydantic_core==2.16.3
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pypdf==4.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
ruamel.yaml==0.17.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.13.0
sentence-transformers==2.6.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
striprtf==0.0.26
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
time-machine==2.14.1
tokenizers==0.15.2
torch==2.1.0
tqdm==4.66.2
transformers==4.38.1
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
